feat: rename config to competition
