+++ @robopo/web/app/competition/loading.tsx
