+++ @robopo/web/app/competition/tabs.tsx
