+++ @robopo/web/app/competition/view.tsx
@@ -10,8 +10,11 @@ import {
   TrashIcon,
 } from "@heroicons/react/24/outline"
 import { useState } from "react"
+import {
+  CompetitionFormModal,
+  DeleteCompetitionModal,
+} from "@/app/competition/tabs"
 import { CommonCheckboxList } from "@/app/components/common/commonList"
-import { CompetitionFormModal, DeleteCompetitionModal } from "@/app/config/tabs"
 import {
   type CompetitionStatus,
   getCompetitionStatus,
